+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - Ramp up missile spawning over time in planetary_defense scenario.
 - Added vector cross product to API.
+- Fixed format code action.
 
 ### 0.39.2